requirements.txt
Flask==2.1.1
pillow==9.0.1
numpy==1.21.6
tensorflow==2.8.0
gunicorn==20.1.0

